requirements.txt
asgiref==3.2.7
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-silk==4.0.1
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
Pygments==2.6.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
wincertstore==0.2